Combine insertion-sort with linked-list
The new implementation, comes from the idea presented in `E 2.3-6` that
suggests that we might be able to combine binary-search with insertion-
sort which is delightful but unattainable (explained with details on
63a759f6d3434a9cbfa6c24fda64eb44263fc7af). However we can use a
linked-list which does not require us to change all the indices if we
were to push an element into it. As pleasing it might sound, if we skip
the indices, then how can we find the middle of the array? I've decided
to keep track of the middle element only. So at least if we were to
perform a linear-search, we only have to search half of the list instead
of all of it.